rename cliview to clicore; more meaningful and consistent
