+++ cli/clicore/examples/basic/typegen.go
